refactor and check that the submit get all the right values
