+++ our-brothers-web/src/app/donate/page/donate-page.module.ts
@@ -1,8 +1,8 @@
 import { NgModule } from '@angular/core';
 import { DonatePageComponent } from './donate-page.component';
 import { RouterModule } from '@angular/router';
-import { PaymentInfoFormComponent } from './payment-info-form/payment-info-form.component';
 import { ReactiveFormsModule } from '@angular/forms';
+import { CommonModule } from '@angular/common';
 
 const routes = [
   {
@@ -12,7 +12,7 @@ const routes = [
 ];
 
 @NgModule({
-  imports: [RouterModule.forChild(routes), ReactiveFormsModule],
-  declarations: [DonatePageComponent, PaymentInfoFormComponent]
+  imports: [RouterModule.forChild(routes), ReactiveFormsModule, CommonModule],
+  declarations: [DonatePageComponent]
 })
 export class DonatePageModule {}
